requirements.txt
codecov==2.1.10
connexion[swagger-ui]==2.7.0
coverage==5.3
dataclasses-json==0.3.6
importlib-resources==3.3.0
Flask==1.1.4
marshmallow-dataclass==7.2.1
pandas==1.3.5
pytest==6.2.5
pytest-cov==3.0.0
tox==3.20.1
